backend/requirements.txt
Flask==3.1.2
Flask-Cors==5.0.0
Pillow==11.3.0
Werkzeug==3.1.3
certifi==2025.8.3
itsdangerous==2.2.0
numpy==2.1.3
pymongo==4.14.1
tensorflow==2.19.1
